Push the bits that change into a config file
